[ALPS02840045] mtk_cpuidle: fix coding style
[Detail]
Change coding style of saving/restoring debug registers

MTK-Commit-Id: 0898204b5d97e2f7872a7c149bcc7abc82a58e6b

CR-Id: ALPS02840045
Feature: Power Management
(cherry picked from commit 7b4510c52765a874729745a00dc6ba5907fcdded)

+++ drivers/misc/mediatek/base/power/cpuidle_v2/mtk_cpuidle.c
@@ -61,7 +61,7 @@ static unsigned int c2k_wdt_irq_nr;
 
 #define CPU_IDLE_STA_OFFSET 10
 
-static unsigned int dbg_data[40];
+static unsigned long dbg_data[40];
 static int mtk_cpuidle_initialized;
 
 static void mtk_spm_irq_set_pending(int wakeup_src, int irq_nr)
